refactor: Move validation into util_schemas
