+++ src/cargo/util_schemas/manifest.rs
@@ -16,8 +16,8 @@ use serde::ser;
 use serde::{Deserialize, Serialize};
 use serde_untagged::UntaggedEnumVisitor;
 
-use crate::util::restricted_names;
 use crate::util_schemas::core::PackageIdSpec;
+use crate::util_schemas::restricted_names;
 use crate::util_semver::PartialVersion;
 
 /// This type is used to deserialize `Cargo.toml` files.
